chore: reorganize project structure
follows instruction from
[Structuring Your Project](https://docs.python-guide.org/writing/structure/) and [Kenneth Reitz' recommendation](https://kennethreitz.org/essays/2013/01/27/repository-structure-and-python). see a example at https://github.com/navdeep-G/samplemod/tree/master
